fix: CHIP8 memory corrected to uint8_t[0x1000] aka 4KB (wrongly set as uint16_t[0x1000])

+++ include/chip8.hpp
@@ -7,7 +7,7 @@
 
 class CHIP8 {
 public:
-  uint16_t memory[0x1000]; // 4kb memory
+  uint8_t memory[0x1000];  // 4kb memory
   CPU cpu;                 // System CPU
   Screen screen;
